server/pyproject.toml
[project]
name = "tts-tax-server"
version = "0.1.0"
description = "TTS Tax App - Django backend for professional tax preparation"
requires-python = ">=3.13,<3.15"
dependencies = [
    "django>=5.2,<6.0",
    "psycopg[binary]>=3.2,<4.0",
    "python-dotenv>=1.0,<2.0",
    "djangorestframework>=3.15,<4.0",
    "openpyxl (>=3.1.5,<4.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest>=8.0,<10.0",
    "pytest-django>=4.9,<6.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["tests.py", "test_*.py"]
pythonpath = ["."]
